common: pull up `param_psbt`
Usable other places, not just in wallet
