+++ tests/test_wallet.py
@@ -789,14 +789,14 @@ def test_sign_and_send_psbt(node_factory, bitcoind, chainparams):
         bitcoind.rpc.sendrawtransaction(broadcast_tx['tx'])
 
     # Try an empty PSBT
-    with pytest.raises(RpcError, match=r"should be a PSBT, not"):
+    with pytest.raises(RpcError, match=r"psbt: Expected a PSBT: invalid token"):
         l1.rpc.signpsbt('')
-    with pytest.raises(RpcError, match=r"should be a PSBT, not"):
+    with pytest.raises(RpcError, match=r"psbt: Expected a PSBT: invalid token"):
         l1.rpc.sendpsbt('')
 
     # Try an invalid PSBT string
     invalid_psbt = 'cHNidP8BAM0CAAAABJ9446mTRp/ml8OxSLC1hEvrcxG1L02AG7YZ4syHon2sAQAAAAD9////JFJH/NjKwjwrP9myuU68G7t8Q4VIChH0KUkZ5hSAyqcAAAAAAP3///8Uhrj0XDRhGRno8V7qEe4hHvZcmEjt3LQSIXWc+QU2tAEAAAAA/f///wstLikuBrgZJI83VPaY8aM7aPe5U6TMb06+jvGYzQLEAQAAAAD9////AcDGLQAAAAAAFgAUyQltQ/QI6lJgICYsza18hRa5KoEAAAAAAAEBH0BCDwAAAAAAFgAUqc1Qh7Q5kY1noDksmj7cJmHaIbQAAQEfQEIPAAAAAAAWABS3bdYeQbXvBSryHNoyYIiMBwu5rwABASBAQg8AAAAAABepFD1r0NuqAA+R7zDiXrlP7J+/PcNZhwEEFgAUKvGgVL/ThjWE/P1oORVXh/ObucYAAQEgQEIPAAAAAAAXqRRsrE5ugA1VJnAith5+msRMUTwl8ocBBBYAFMrfGCiLi0ZnOCY83ERKJ1sLYMY8A='
-    with pytest.raises(RpcError, match=r"should be a PSBT, not"):
+    with pytest.raises(RpcError, match=r"psbt: Expected a PSBT: invalid token"):
         l1.rpc.signpsbt(invalid_psbt)
 
     wallet_coin_mvts = [
